requirements.txt
Django==6.0.2
djangorestframework==3.16.1
django-cors-headers==4.9.0
google-generativeai==0.8.6
python-decouple==3.8